add SHA256SUMS to the artifacts
As of now, github does include checksums for assets: https://github.blog/changelog/2025-06-03-releases-now-expose-digests-for-release-assets/

However, it's locked behind github's API, which is cumbersome to use
and rate-limited to oblivion without an API token.

Having a checksum file included in the artifacts
is likely to remain the best way to acquire the checksums.